improvements
- Support parent processes matching
- Add Jest
- Refactor folder structure
