+++ scripts/dl_geosite.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 
-curl -o helper/geosite.dat -L https://github.com/v2ray/domain-list-community/releases/latest/download/dlc.dat
+curl -o src/helper/geosite.dat -L https://github.com/v2ray/domain-list-community/releases/latest/download/dlc.dat
