+++ scripts/dl_mmdb.sh
@@ -4,4 +4,4 @@ curl -o /tmp/mmdb.tar.gz -L http://geolite.maxmind.com/download/geoip/database/G
 rm -rf /tmp/mmdb > /dev/null 2>&1
 mkdir -p /tmp/mmdb
 tar -xf /tmp/mmdb.tar.gz -C /tmp/mmdb --strip-components 1
-cp /tmp/mmdb/GeoLite2-Country.mmdb helper/geo.mmdb
+cp /tmp/mmdb/GeoLite2-Country.mmdb src/helper/geo.mmdb
