+++ src/spec/null.spec.js
@@ -1,0 +1,5 @@
+describe('Dummy test', () => {
+  test('Should just pass', () => {
+    expect(true).toEqual(true);
+  })
+})
